requirements.txt
attrs==16.3.0
cffi==1.9.1
constantly==15.1.0
cryptography==1.7.1
cssselect==1.0.0
Django==1.10.5
django-filter==1.0.1
djangorestframework==3.5.3
enum34==1.1.6
idna==2.2
incremental==16.10.1
ipaddress==1.0.17
lxml==3.7.1
Markdown==2.6.7
mysql==0.0.1
MySQL-python==1.2.5
parsel==1.1.0
poster==0.8.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
pyOpenSSL==16.2.0
queuelib==1.4.2
requests==2.12.4
Scrapy==1.3.0
service-identity==16.0.0
six==1.10.0
Twisted==16.6.0
Unirest==1.1.7
w3lib==1.16.0
zope.interface==4.3.3
